pyproject.toml
# File generated by saltoapis-python. DO NOT EDIT.
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "saltoapis-python"
dynamic = ["version"]
description = "Saltoapis Python SDK"
dependencies = [
  'grpcio==1.63.0',
  'grpcio-tools==1.63.0',
  'protobuf==5.26.1',
]
license = {file = "LICENSE"}

[tool.setuptools]
packages = [
    'saltoapis.nebula.accesspointgroup.v1',
    'saltoapis.nebula.file.v1',
    'saltoapis.nebula.iam.v1',
    'saltoapis.nebula.unit.v1',
    'saltoapis.type',
    'google.rpc',
    'saltoapis.nebula.digitalkey.v1',
    'saltoapis.nebula.installation.v1',
    'saltoapis.nebula.type',
    'saltoapis.longrunning.v1',
    'saltoapis.nebula.accesspoint.v1',
    'saltoapis.nebula.accessright.v1',
    'saltoapis.nebula.calendar.v1',
    'saltoapis.nebula.controller.v1',
    'saltoapis.nebula.device.v1',
    'saltoapis.nebula.electroniclock.v1',
    'saltoapis.nebula.encoder.v1',
    'saltoapis.nebula.extender.v1',
    'saltoapis.nebula.gateway.v1',
    'saltoapis.nebula.intercomadaptor.v1',
    'saltoapis.nebula.openingmodeschedule.v1',
    'saltoapis.nebula.user.v1',
    'saltoapis.nebula.event.v1',
    'saltoapis.auth',
]
